requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.1
astroid==2.4.0
astropy==3.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.4
bokeh==1.0.2
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.0.0
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
comtypes==1.1.7
contextlib2==0.6.0.post1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==1.0.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed==1.28.1
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-Cors==3.0.8
fsspec==0.7.1
future==0.18.2
gast==0.2.2
gevent==1.4.0
glob2==0.7
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hypothesis==5.11.0
idna==2.9
imageio==2.4.1
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.26.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.41.0
numexpr==2.6.8
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.2.0
packaging==20.3
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.7.0
PuLP==1.6.10
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.5
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.9
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.2.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
PyWavelets==1.0.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.0
ruamel-yaml==0.15.87
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.16
statsmodels==0.9.0
sympy==1.5.1
tables==3.4.4
tblib==1.6.0
tensorboard==2.1.1
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.19.0
xlwt==1.3.0
yapf==0.28.0
zict==2.0.0
zipp==3.1.0
